src/requirements.txt
flask==2.3.0
mysql-connector-python==8.1.0